Jiaruifang/cpu variable len benchmark (#142)
* Reorganized benchmark scripts
